Move SerializeableRequest request to remote_attestation_sessions to minimze dependencies of the channel crate
