+++ experimental/uefi/channel/Cargo.toml
@@ -5,10 +5,6 @@ authors = ["Juliette Pretot <julsh@google.com>"]
 edition = "2021"
 license = "Apache-2.0"
 
-[features]
-default = ["ring-crypto"]
-ring-crypto = ["oak_remote_attestation_sessions/ring-crypto"]
-
 [dependencies]
 anyhow = { version = "*", default-features = false }
 oak_remote_attestation_sessions = { path = "../../../remote_attestation_sessions", default-features = false }
