+++ experimental/uefi/channel/src/lib.rs
@@ -78,47 +78,3 @@ where
         Ok(())
     }
 }
-
-pub struct SerializeableRequest {
-    pub session_id: SessionId,
-    pub body: Vec<u8>,
-}
-
-impl From<SerializeableRequest> for Vec<u8> {
-    fn from(serializeable_request: SerializeableRequest) -> Vec<u8> {
-        // The payload is the request's body prepended with the 8 byte session_id.
-        // This takes adavantage of the session_id's fixed size to avoid needing
-        // to use a key/value pair binary serialization protocol.
-        let mut serialized_request: Vec<u8> = Vec::with_capacity(
-            serializeable_request.session_id.len() + serializeable_request.body.len(),
-        );
-
-        serialized_request.extend(serializeable_request.session_id);
-        serialized_request.extend(serializeable_request.body);
-
-        serialized_request
-    }
-}
-
-impl TryFrom<&[u8]> for SerializeableRequest {
-    type Error = anyhow::Error;
-
-    fn try_from(serialized_request: &[u8]) -> Result<Self, Self::Error> {
-        if serialized_request.len() < SESSION_ID_LENGTH {
-            bail!(
-                "Message too short to contain a SessionId. The length of a SessionId
-                is {} bytes, the message received contained only {} bytes",
-                SESSION_ID_LENGTH,
-                serialized_request.len()
-            );
-        }
-
-        let (session_id_slice, request_body_slice) = serialized_request.split_at(SESSION_ID_LENGTH);
-
-        let mut session_id: SessionId = [0; SESSION_ID_LENGTH];
-        session_id.copy_from_slice(session_id_slice);
-        let body = request_body_slice.to_vec();
-
-        Ok(Self { session_id, body })
-    }
-}
