+++ experimental/uefi/runtime/src/remote_attestation.rs
@@ -23,8 +23,7 @@ extern crate alloc;
 
 use alloc::vec::Vec;
 use anyhow::Context;
-use channel::SerializeableRequest;
-use oak_remote_attestation_sessions::{SessionState, SessionTracker};
+use oak_remote_attestation_sessions::{SerializeableRequest, SessionState, SessionTracker};
 
 /// Number of sessions that will be kept in memory.
 const SESSIONS_CACHE_SIZE: usize = 10000;
